added some more code for controller

+++ bms/src/main/java/com/king/bms/dao/MovieDao.java
@@ -13,17 +13,8 @@ public interface MovieDao extends JpaRepository<Movie, Integer> {
     // to find movie by movie_name;
     public Movie findMovieByMovieName(String movieName);
 
-
-    // to find movies by city
-    //@Query("select Theaters from Movie where name=:name")
-   // public List<Theater> findAllTheaterByMovieName(@Param("name") String name);
-
-    //public List<Theater> findAllTheaterByMovieName(String name);
-
-
-
     //I want to search based on conditions
-    public List<Movie> findByDurationGreaterThanEqual(int duration);
+    public List<Movie> findByDurationGreaterThanEqual(float duration);
 
     public List<Movie> findByMovieNameContaining(String movieString);
